refactor: refactored the correlation-ids middleware
Closes #55
