+++ packages/lambda-powertools-middleware-correlation-ids/consts.js
@@ -1,0 +1,6 @@
+module.exports = {
+  X_CORRELATION_ID: 'x-correlation-id',
+  DEBUG_LOG_ENABLED: 'debug-log-enabled',
+  USER_AGENT: 'User-Agent',
+  CALL_CHAIN_LENGTH: 'call-chain-length'
+}
